Reject tweets which where retweeted from the user we follow
Detailed explanation:

The user we follow sends a tweet.
Someone retweets the tweet.
We somehow (dafuq? Streamingapi) get this retweet.
We should not try to retweet the retweet because we already retweeted the tweet and the API will block our attempt to retweet the retweet. This block is currently handled by crashing causing a retry on next start leading to a new crash